requirements.txt
certifi==2017.11.5
click==6.7
Flask==0.12.1
future==0.16.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pygraphviz==1.3.1
python-telegram-bot==5.3.0
six==1.11.0
transitions==0.5.0
urllib3==1.22
Werkzeug==0.13
